requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
decorator==5.2.1
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
idna==3.10
iniconfig==2.1.0
jellyfin-apiclient-python==1.11.0
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.1
multidict==6.6.4
numba==0.61.2
numpy==2.2.6
opuslib==3.0.1
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
propcache==0.3.2
protobuf==6.32.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pymumble @ git+https://github.com/DuckBoss/pymumble@cefea66a86bc3eb0a80943697a61bd210379476a
pytest==8.4.1
pyttsx3==2.99
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
yarl==1.20.1
